requirements.txt
rich
asyncio
loguru
web3
pathlib
urllib3
pytest-playwright